created a separate file for the queries at lambda_model.js
